+++ shaker_sort.h
@@ -2,6 +2,6 @@
 #define SHAKER_SORT_H_
 
 void shakerSort(int a[], int n);
-int shakerSortComparisions(int a[], int n);
+long long shakerSortComparisions(int a[], int n);
 
 #endif
